Try including gcov and increased static analysis

+++ src/string/tokenizer.c
@@ -38,7 +38,7 @@ char **tokenize_string(const char *s, const char *delimiter, size_t *num_tokens)
 }
 
 void free_array_of_strings(char **tokens, size_t number_of_tokens) {
-    for (int i = 0; i < number_of_tokens; i++) {
+    for (size_t i = 0; i < number_of_tokens; i++) {
         free(tokens[i]);
     }
     free(tokens);
